Moving document mapping endpoints out of EndpointContracts folder
