--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/IRazorMapToDocumentEditsHandler.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Razor.LanguageServer.Common;
-using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
-using Microsoft.CommonLanguageServerProtocol.Framework;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
-
-[LanguageServerEndpoint(LanguageServerConstants.RazorMapToDocumentEditsEndpoint)]
-internal interface IRazorMapToDocumentEditsHandler : IRazorRequestHandler<RazorMapToDocumentEditsParams, RazorMapToDocumentEditsResponse>
-{
-}
