+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Mapping/RazorMapToDocumentRangesEndpoint.cs
@@ -5,13 +5,18 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.LanguageServer.Common;
 using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+using Microsoft.CommonLanguageServerProtocol.Framework;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Mapping;
 
-internal class RazorMapToDocumentRangesEndpoint : IRazorMapToDocumentRangesHandler
+[LanguageServerEndpoint(LanguageServerConstants.RazorMapToDocumentRangesEndpoint)]
+internal sealed class RazorMapToDocumentRangesEndpoint :
+    IRazorDocumentlessRequestHandler<RazorMapToDocumentRangesParams, RazorMapToDocumentRangesResponse?>,
+    ITextDocumentIdentifierHandler<RazorMapToDocumentRangesParams, Uri>
 {
     private readonly IRazorDocumentMappingService _documentMappingService;
