+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Mapping/RazorMapToDocumentRangesParams.cs
@@ -5,7 +5,7 @@ using System;
 using System.Runtime.Serialization;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Mapping;
 
 // Note: This type should be kept in sync with the one in VisualStudio.LanguageServerClient assembly.
 [DataContract]
