+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Mapping/RazorMapToDocumentRangesEndpointTest.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.CodeGeneration;
-using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using Microsoft.AspNetCore.Razor.LanguageServer.Test.Common;
@@ -17,13 +16,13 @@ using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Mapping;
 
-public class RazorLanguageEndpointTest : LanguageServerTestBase
+public class RazorMapToDocumentRangesEndpointTest : LanguageServerTestBase
 {
     private readonly IRazorDocumentMappingService _mappingService;
 
-    public RazorLanguageEndpointTest(ITestOutputHelper testOutput)
+    public RazorMapToDocumentRangesEndpointTest(ITestOutputHelper testOutput)
         : base(testOutput)
     {
         _mappingService = new RazorDocumentMappingService(
@@ -249,117 +248,6 @@ public class RazorLanguageEndpointTest : LanguageServerTestBase
         // Assert
         Assert.NotNull(response);
         Assert.Equal(RangeExtensions.UndefinedRange, response!.Ranges[0]);
-        Assert.Equal(1337, response.HostDocumentVersion);
-    }
-
-    [Fact]
-    public async Task Handle_ResolvesLanguageRequest_Razor()
-    {
-        // Arrange
-        var documentPath = new Uri("C:/path/to/document.cshtml");
-        var codeDocument = CreateCodeDocument("@{}");
-        var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var languageEndpoint = new RazorLanguageQueryEndpoint(_mappingService);
-        var request = new RazorLanguageQueryParams()
-        {
-            Uri = documentPath,
-            Position = new Position(0, 1),
-        };
-
-        var requestContext = CreateRazorRequestContext(documentContext);
-
-        // Act
-        var response = await languageEndpoint.HandleRequestAsync(request, requestContext, default);
-
-        // Assert
-        Assert.Equal(RazorLanguageKind.Razor, response.Kind);
-        Assert.Equal(request.Position, response.Position);
-        Assert.Equal(1337, response.HostDocumentVersion);
-    }
-
-    // This is more of an integration test to validate that all the pieces work together
-    [Fact]
-    public async Task Handle_ResolvesLanguageRequest_Html()
-    {
-        // Arrange
-        var documentPath = new Uri("C:/path/to/document.cshtml");
-        var codeDocument = CreateCodeDocument("<s");
-        var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var languageEndpoint = new RazorLanguageQueryEndpoint(_mappingService);
-        var request = new RazorLanguageQueryParams()
-        {
-            Uri = documentPath,
-            Position = new Position(0, 2),
-        };
-
-        var requestContext = CreateRazorRequestContext(documentContext);
-
-        // Act
-        var response = await languageEndpoint.HandleRequestAsync(request, requestContext, default);
-
-        // Assert
-        Assert.Equal(RazorLanguageKind.Html, response.Kind);
-        Assert.Equal(request.Position, response.Position);
-        Assert.Equal(1337, response.HostDocumentVersion);
-    }
-
-    // This is more of an integration test to validate that all the pieces work together
-    [Fact]
-    public async Task Handle_ResolvesLanguageRequest_CSharp()
-    {
-        // Arrange
-        var documentPath = new Uri("C:/path/to/document.cshtml");
-        var codeDocument = CreateCodeDocumentWithCSharpProjection(
-            "@",
-            "/* CSharp */",
-            new[] { new SourceMapping(new SourceSpan(0, 1), new SourceSpan(0, 12)) });
-        var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var languageEndpoint = new RazorLanguageQueryEndpoint(_mappingService);
-        var request = new RazorLanguageQueryParams()
-        {
-            Uri = documentPath,
-            Position = new Position(0, 1),
-        };
-        var requestContext = CreateRazorRequestContext(documentContext);
-
-        // Act
-        var response = await languageEndpoint.HandleRequestAsync(request, requestContext, default);
-
-        // Assert
-        Assert.Equal(RazorLanguageKind.CSharp, response.Kind);
-        Assert.Equal(0, response.Position.Line);
-        Assert.Equal(1, response.Position.Character);
-        Assert.Equal(1337, response.HostDocumentVersion);
-    }
-
-    // This is more of an integration test to validate that all the pieces work together
-    [Fact]
-    public async Task Handle_Unsupported_ResolvesLanguageRequest_Html()
-    {
-        // Arrange
-        var documentPath = new Uri("C:/path/to/document.cshtml");
-        var codeDocument = CreateCodeDocumentWithCSharpProjection(
-            "@",
-            "/* CSharp */",
-            new[] { new SourceMapping(new SourceSpan(0, 1), new SourceSpan(0, 12)) });
-        codeDocument.SetUnsupported();
-        var documentContext = CreateDocumentContext(documentPath, codeDocument);
-        var languageEndpoint = new RazorLanguageQueryEndpoint(_mappingService);
-        var request = new RazorLanguageQueryParams()
-        {
-            Uri = documentPath,
-            Position = new Position(0, 1),
-        };
-
-        var requestContext = CreateRazorRequestContext(documentContext);
-
-        // Act
-        var response = await languageEndpoint.HandleRequestAsync(request, requestContext, default);
-
-        // Assert
-        Assert.Equal(RazorLanguageKind.Html, response.Kind);
-        Assert.Equal(0, response.Position.Line);
-        Assert.Equal(1, response.Position.Character);
         Assert.Equal(1337, response.HostDocumentVersion);
     }
